Print out full section stacks on failure.
The Console reporter now "traces" the full section stack of its
current test execution and prints it out instead of the section
stack of the failing assertion itself. Likewise, the JUnit reporter
now groups assertions together into the <testcase> of the test
execution where they happened.

This fixes #734 and allows post-condition tests after the deepest
SECTION in a test execution.